+++ backend/src/main/java/com/ssafy/jupging/domain/repository/ArticleRepository.java
@@ -10,9 +10,9 @@ import java.util.Optional;
 
 public interface ArticleRepository extends JpaRepository<Article, Long> {
 
-    Optional<Article> findByArticleId(int article_id);
+    Optional<Article> findByArticleId(Long article_id);
 
-    List<Article> findByUserId(int user_id);
+    List<Article> findByUserId(Long user_id);
 
     List<Article> findTop10ByOrderByCreatedDateDesc();
 }
